refactor(auth): duplicate JWT auth in app and api folders
